refactor: Move process handling code to pueue_lib
