+++ Cargo.toml
@@ -33,7 +33,7 @@ name = "pueue_daemon_lib"
 path = "daemon/lib.rs"
 
 [dependencies]
-pueue-lib = { version = "0.20.0", path = "lib" }
+pueue-lib = { version = "0.21.0", path = "lib" }
 
 anyhow = "1"
 chrono = { version = "0.4", features = ["serde"] }
@@ -55,28 +55,15 @@ snap = "1"
 tempfile = "3"
 tokio = { version = "1", features = ["rt-multi-thread", "time", "io-std"] }
 
-# Windows
-[target.'cfg(windows)'.dependencies]
-winapi = { version = "0.3", features = [
-  "tlhelp32",
-  "handleapi",
-  "processthreadsapi",
-  "minwindef",
-  "impl-default",
-] }
-
-# Unix
-[target.'cfg(not(windows))'.dependencies]
-nix = "0.24"
-whoami = "1"
-
-# Linux / Freebsd
-[target.'cfg(any(target_os = "linux", target_os = "freebsd"))'.dependencies]
-procfs = { version = "0.12", default-features = false }
-
 [dev-dependencies]
 anyhow = "1"
 assert_cmd = "2"
 better-panic = "0.3"
 pretty_assertions = "1"
 rstest = "0.15"
+
+# Test specific dev-dependencies
+[target.'cfg(any(target_os = "linux", target_os = "freebsd"))'.dependencies]
+nix = "0.24"
+whoami = "1"
+procfs = { version = "0.12", default-features = false }
