--- daemon/platform/apple/mod.rs
@@ -1,1 +1,0 @@
-pub mod process_helper;
