+++ daemon/task_handler/messages/kill.rs
@@ -1,11 +1,11 @@
 use log::{error, info, warn};
 
 use pueue_lib::network::message::{Signal, TaskSelection};
+use pueue_lib::process_helper::*;
 use pueue_lib::state::GroupStatus;
 use pueue_lib::task::TaskStatus;
 
 use crate::ok_or_shutdown;
-use crate::platform::process_helper::*;
 use crate::state_helper::save_state;
 use crate::task_handler::{Shutdown, TaskHandler};
