+++ lib/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "pueue-lib"
 description = "The shared library to work with the Pueue client and daemon."
-version = "0.20.0"
+version = "0.21.0"
 authors = ["Arne Beer <contact@arne.beer>"]
 homepage = "https://github.com/nukesor/pueue"
 repository = "https://github.com/nukesor/pueue"
@@ -16,7 +16,7 @@ edition = "2021"
 maintenance = { status = "actively-developed" }
 
 [dependencies]
-
+anyhow = "1"
 async-trait = "0.1"
 byteorder = "1"
 chrono = { version = "0.4", features = ["serde"] }
@@ -45,9 +45,6 @@ tokio = { version = "1", features = [
 ] }
 tokio-rustls = "0.23"
 
-[target.'cfg(not(windows))'.dependencies]
-whoami = "1"
-
 [dev-dependencies]
 anyhow = "1"
 better-panic = "0.3"
@@ -55,3 +52,23 @@ pretty_assertions = "1"
 portpicker = "0.1"
 tempdir = "0.3"
 tokio = { version = "1", features = [ "time" ] }
+
+# --- Platform specific dependencies ---
+# Windows
+[target.'cfg(windows)'.dependencies]
+winapi = { version = "0.3", features = [
+  "tlhelp32",
+  "handleapi",
+  "processthreadsapi",
+  "minwindef",
+  "impl-default",
+] }
+
+# Unix
+[target.'cfg(not(windows))'.dependencies]
+nix = "0.24"
+whoami = "1"
+
+# Linux / Freebsd
+[target.'cfg(any(target_os = "linux", target_os = "freebsd"))'.dependencies]
+procfs = { version = "0.12", default-features = false }
